refactor(tests/unit/utils.py): Move sharing helper function to utils.py
