+++ tests/unit/commands/conftest.py
@@ -1,4 +1,3 @@
-from typing import AsyncGenerator, List
 from unittest.mock import AsyncMock
 
 import pytest
@@ -16,9 +15,3 @@ def test_juju_command():
     command = TestJujuCommand()
     command.execute.reset_mock()
     yield command
-
-
-async def _async_generator(values: List) -> AsyncGenerator:
-    """Async generator."""
-    for value in values:
-        yield value
